+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Mermaid 9.2.2 support
 - Chart.js Plugin matrix 1.2.1 support
+- D3 7.7.0 support
 
 ## 3.0.7