+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.0.8
 
-- Mermaid 9.2.2 support
+- Mermaid 9.3.0 support
 - Chart.js Plugin matrix 1.2.1 support
 - D3 7.7.0 support